Update benchmarks with full 15-video run results
PE-Core L14: 7.00 fps avg (fastest)
Jina v4 Single Frame: 0.19 fps avg (batch_size=1 optimal on MPS)
Jina v4 Grid Composite: 0.73 fps avg
Full run: 9h50m across 5,764 frames

Also: per-frame progress logging for Jina, torch.mps.synchronize()
between pipeline groups to prevent MPS hangs, reverted Jina to float32
(float16 caused hangs via sentence-transformers), removed SigLIP from
active pipelines.

+++ app/pipelines/jina_native3d.py
@@ -37,7 +37,7 @@ class JinaNative3DPipeline(BaseEmbeddingPipeline):
             AutoModel.from_pretrained(
                 self.MODEL_ID,
                 trust_remote_code=True,
-                dtype=torch.float16,
+                dtype=torch.float32,
             )
             .to(self.device)
             .eval()
